requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LLM integration
litellm>=0.12.0
openai>=1.2.4

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Utilities
tenacity>=8.2.3
uuid>=1.30
python-dateutil>=2.8.2

# CSV handling
pandas>=2.1.3

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.7 